ensure evrything is clickable and working
