+++ lib/lightning_web/live/credential_live/credential_index_component.html.heex
@@ -126,7 +126,7 @@
     id="new-credential-modal"
     module={LightningWeb.CredentialLive.CredentialFormComponent}
     action={:new}
-    credential_type={@selected_credential_type}
+    credential_type={nil}
     credential={@credential}
     oauth_client={nil}
     oauth_clients={@oauth_clients}
@@ -141,7 +141,7 @@
     id={"edit-credential-#{@credential.id}-modal"}
     module={LightningWeb.CredentialLive.CredentialFormComponent}
     action={:edit}
-    credential_type={@selected_credential_type}
+    credential_type={nil}
     credential={@credential}
     oauth_client={@oauth_client}
     project={@project}
