+++ lib/lightning_web/live/job_live/credential_picker.ex
@@ -7,7 +7,6 @@ defmodule LightningWeb.JobLive.CredentialPicker do
   use LightningWeb, :live_component
 
   alias LightningWeb.Components.Form
-  alias Phoenix.LiveView.JS
 
   attr :form, :map, required: true
   attr :disabled, :boolean, default: false
@@ -45,50 +44,11 @@ defmodule LightningWeb.JobLive.CredentialPicker do
           id="new-credential-button"
           type="button"
           class="link text-xs"
-          phx-click="toggle_credentials_modal"
-          phx-target={@myself}
+          phx-click="toggle_job_credential_modal"
         >
           New credential
         </button>
       </div>
-
-      <.live_component
-        :if={!@disabled and @show_credential_modal}
-        id="new-credential-modal"
-        module={LightningWeb.CredentialLive.CredentialFormComponent}
-        action={:new}
-        credential_type={nil}
-        credential={
-          %Lightning.Credentials.Credential{
-            user_id: @current_user.id,
-            project_credentials: [
-              %Lightning.Projects.ProjectCredential{
-                project_id: @project.id
-              }
-            ]
-          }
-        }
-        current_user={@current_user}
-        oauth_client={nil}
-        oauth_clients={@oauth_clients}
-        projects={[]}
-        project={@project}
-        on_save={
-          fn credential ->
-            params =
-              LightningWeb.Utils.build_params_for_field(
-                @form,
-                :project_credential_id,
-                credential.project_credentials |> Enum.at(0) |> Map.get(:id)
-              )
-
-            @on_change.(params)
-          end
-        }
-        on_modal_close={JS.push("toggle_credentials_modal", target: @myself)}
-        can_create_project_credential={!@disabled}
-        return_to={@credential_modal_return_to}
-      />
     </div>
     """
   end
@@ -99,8 +59,7 @@ defmodule LightningWeb.JobLive.CredentialPicker do
      socket
      |> assign(
        credentials: [],
-       selected_project_credential_id: nil,
-       show_credential_modal: false
+       selected_project_credential_id: nil
      )}
   end
 
@@ -134,10 +93,5 @@ defmodule LightningWeb.JobLive.CredentialPicker do
       end)
 
     {:ok, socket}
-  end
-
-  @impl true
-  def handle_event("toggle_credentials_modal", _params, socket) do
-    {:noreply, update(socket, :show_credential_modal, fn show -> !show end)}
   end
 end
