+++ lib/lightning_web/live/project_live/settings.html.heex
@@ -253,12 +253,12 @@
             %{
               name: "Credential",
               id: "option-menu-item-1",
-              target: "new-credential-modal"
+              target: "new_credential"
             },
             %{
               name: "OAuth client",
               id: "option-menu-item-0",
-              target: "new-oauth-client-modal",
+              target: "new_oauth_client",
               badge: "Advanced"
             }
           ]}
